requirements.txt
discord.py==2.5.2
python-dotenv==1.1.1
Pillow==10.4.0
geopy==2.4.1